Data Wrangling
Assignment Overview
Data Wrangling or Munging is a process in which we clean up the data set and make it ready for data analysis. In this assignment you will perform the following tasks:

Identify duplicate rows in the data frame.

Remove duplicate rows from the dataframe.

Find the number of missing values for all columns.

Find the value counts for the column "Employment".

Normalize the data using two existing columns.